backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.34
asyncmy==0.2.9
alembic==1.13.2
pymysql==1.1.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9

openpyxl==3.1.5
httpx==0.27.2
email-validator==2.1.1
bcrypt==4.0.1

